// src/app/models/stolper-stein.model.ts
export interface StolperStein {
    id: string;
    firstName: string;
    preposition: string;
    lastName: string;
    street: string;
    houseNumber: string;
    city: string;
    birthyear: number;
    dateOfDeath: any;
    location: string;
    dateOfStolperSteinLaying: any;
    weblink: string;
    latitude: number;
    longitude: number;
    distanceFromUserInMeters: number
}